+++ balances/index.js
@@ -1,10 +1,12 @@
 const accountingCategories = require('./accounting_categories');
+const balanceFromTokens = require('./balance_from_tokens');
 const getAccountingReport = require('./get_accounting_report');
 const getBalance = require('./get_balance');
 const getLiquidity = require('./get_liquidity');
 
 module.exports = {
   accountingCategories,
+  balanceFromTokens,
   getAccountingReport,
   getBalance,
   getLiquidity,