+++ responses/return_number.js
@@ -15,7 +15,9 @@
 module.exports = ({logger, number, reject, resolve}) => {
   return (err, res) => {
     if (!!err) {
-      return reject(err);
+      logger.error(err);
+
+      return reject();
     }
 
     logger.info(`${res[number]}`);