+++ swaps/index.js
@@ -1,5 +1,15 @@
+const getSwapCost = require('./get_swap_cost');
+const getSwapService = require('./get_swap_service');
 const rebalance = require('./rebalance');
 const swapIn = require('./swap_in');
 const swapOut = require('./swap_out');
+const {swapTypes} = require('./constants');
 
-module.exports = {rebalance, swapIn, swapOut};
+module.exports = {
+  getSwapCost,
+  getSwapService,
+  rebalance,
+  swapIn,
+  swapOut,
+  swapTypes,
+};