+++ test/responses/test_return_output.js
@@ -17,11 +17,16 @@ const tests = [
 tests.forEach(({args, description, error, expected}) => {
   return test(description, ({end, equal, throws}) => {
     if (!!error) {
-      return returnOutput({reject: err => {
-        equal(err, error, 'Error as expected');
+      let err;
 
-        return end();
-      }})(error);
+      return returnOutput({
+        logger: {error: n => err = n},
+        reject: () => {
+          equal(err, error, 'Error as expected');
+
+          return end();
+        },
+      })(error);
     }
 
     let output;